feat: inverted index by comics base
